+++ app/code/local/Autocompleteplus/Autosuggest/Model/Email/Template.php
@@ -76,53 +76,6 @@ class Autocompleteplus_Autosuggest_Model_Email_Template extends Mage_Core_Model_
     {
         $processedResult = parent::getProcessedTemplate($variables);
 
-        if (!isset($variables['this']) && isset($variables['order'])) {
-            try {
-                $items = array();
-                foreach ($variables['order']->getItemsCollection() as $item) {
-                    $items[] = array(
-                        'product_id' => $item->getProduct()->getId(),
-                        'price' => $item->getrow_total(),
-                        'quantity' => $item->getqty_ordered(),
-                        'currency' => $variables['order']->getorder_currency_code(),
-                        'attribution' => $item->getAddedFromSearch(),
-                    );
-                }
-
-                $items_json = json_encode($items);
-
-                $params = array(
-                    'template' => $processedResult,
-                    'event' => 'email_confirmation',
-                    'UUID' => $this->getConfig()->getUUID(),
-                    'key' => $this->getConfig()->getAuthorizationKey(),
-                    'store_id' => $variables['order']->getStoreId(),
-                    'st' => Mage::helper('autocompleteplus_autosuggest')
-                        ->getSessionId(),
-                    'cart_token' => $variables['order']->getQuoteId(),
-                    'cart_product' => $items_json
-                );
-
-                // @codingStandardsIgnoreStart
-                /**
-                 * Due to backward compatibility issues with Magento < 1.8.1 and cURL/Zend
-                 * We need to use PHP's implementation of cURL directly rather than Zend or Varien.
-                 */
-                $client = curl_init(static::AUTOCOMPLETEPLUS_WEBHOOK_URI);
-                curl_setopt($client, CURLOPT_RETURNTRANSFER, true);
-                curl_setopt($client, CURLOPT_POSTFIELDS, $params);
-                $response = curl_exec($client);
-                curl_close($client);
-            } catch (Exception $e) {
-                Mage::log(
-                    __FILE__.' throws:'.$e->getMessage(),
-                    null,
-                    'autocomplete.log',
-                    true
-                );
-            }
-        }
-
         return $processedResult;
     }
 }